Move page gallery item to Galleries/templates
