+++ resources/assets/components/utilities/Gallery/templates/PageGalleryItem/page-gallery-item.scss
@@ -1,0 +1,7 @@
+@import '../../../../../scss/next-toolbox.scss';
+
+.page-gallery-item {
+  .description {
+    font-weight: $weight-normal;
+  }
+}
